MSRInfrastructure/requirements.txt
gql==3.4.1
pygithub==2.1.1
requests-toolbelt==1.0.0
gitpython==3.1.40
neo4j==5.14.0
python-magic==0.4.27